@@ -2,38 +2,39 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using BenchmarkDotNet.Attributes;
 using DiffPlex;
 using DiffPlex.DiffBuilder;
+using DiffPlex.DiffBuilder.Model;
 
 namespace Perf.DiffPlex
 {
-    internal class DiffPerfTester
+    [MemoryDiagnoser]
+    public class SideBySideDiffBuilderBenchmark
     {
-        private static readonly Random random = new Random();
-        private readonly SideBySideDiffBuilder sideBySideDiffer;
+        private static readonly Random Random = new Random();
+        private readonly SideBySideDiffBuilder sideBySideDiffer = new SideBySideDiffBuilder(new Differ());
+        private string oldText;
+        private string newText;
         private const int MaxLineLength = 150;
         private const double DifferenceAmount = 0.2;
         private const int MaxLines = 8000;
 
-        public DiffPerfTester()
-        {
-            Console.WriteLine("Max number of lines: {0}", MaxLines);
-            Console.WriteLine("Max length of lines: {0}", MaxLineLength);
-            Console.WriteLine("Max difference amount: {0}", DifferenceAmount);
-            Console.WriteLine();
-            sideBySideDiffer = new SideBySideDiffBuilder(new Differ());
-        }
-
-        public void Run()
+        [GlobalSetup]
+        public void SetUp()
         {
             var oldLines = GenerateLines(MaxLines);
             var newLines = MakeDifferent(oldLines, DifferenceAmount);
-            var oldText = Implode(oldLines, Environment.NewLine);
-            var newText = Implode(newLines, Environment.NewLine);
-            new PerfTester().Run(() => sideBySideDiffer.BuildDiffModel(oldText, newText));
+            oldText = Implode(oldLines, Environment.NewLine);
+            newText = Implode(newLines, Environment.NewLine);
         }
 
-
+        [Benchmark]
+        public SideBySideDiffModel BuildDiffModel()
+        {
+            return sideBySideDiffer.BuildDiffModel(oldText, newText);
+        }
+        
         private static string Implode<T>(IEnumerable<T> enumerable, string delim)
         {
             if (enumerable == null) throw new ArgumentNullException(nameof(enumerable));
@@ -49,8 +50,8 @@ namespace Perf.DiffPlex
 
             if (loopedAtLeaseOnce)
                 return result.Remove(result.Length - delim.Length, delim.Length).ToString();
-            else
-                return String.Empty;
+            
+            return String.Empty;
         }
 
         private static IList<string> MakeDifferent(IList<string> lines, double differenceAmount)
@@ -58,10 +59,10 @@ namespace Perf.DiffPlex
             var newLines = new List<string>();
             foreach (var i in Enumerable.Range(0, lines.Count))
             {
-                if(random.NextDouble() <= differenceAmount)
+                if(Random.NextDouble() <= differenceAmount)
                 {
                     // Either delete line or add different one
-                    if(random.Next(2) % 2 == 1)
+                    if(Random.Next(2) % 2 == 1)
                     {
                         newLines.Add(RandomString(MaxLineLength));
                     }
@@ -84,17 +85,16 @@ namespace Perf.DiffPlex
         private static string RandomString(int maxLength)
         {
             var builder = new StringBuilder();
-            foreach (var i in Enumerable.Range(0, random.Next(0, maxLength)))
+            foreach (var i in Enumerable.Range(0, Random.Next(0, maxLength)))
             {
-                var ch = Convert.ToChar(Convert.ToInt32(Math.Floor(26 * random.NextDouble() + 65)));
+                var ch = Convert.ToChar(Convert.ToInt32(Math.Floor(26 * Random.NextDouble() + 65)));
                 builder.Append(ch);
             }
 
-            if (random.Next(2) % 2 == 0)
+            if (Random.Next(2) % 2 == 0)
                 return builder.ToString().ToLower();
 
             return builder.ToString();
         }
-
     }
 }
